requirements.txt
python-dotenv
langgraph
langchain-openai
pydantic
langgraph-cli[inmem]